move BlockNode into its own file
