+++ consensus/state.go
@@ -94,24 +94,6 @@ type MissedStorageProof struct {
 	ContractID ContractID
 }
 
-// A BlockNode contains a block and the list of children to the block. Also
-// contains some consensus information like which contracts have terminated and
-// where there were missed storage proofs.
-type BlockNode struct {
-	Block    Block
-	Children []*BlockNode
-
-	Height           BlockHeight
-	Depth            Target        // What the target would need to be to have a weight equal to all blocks up to this block.
-	Target           Target        // Target for next block.
-	RecentTimestamps [11]Timestamp // The 11 recent timestamps.
-
-	BlockDiff            BlockDiff       // Soon to replace the other 3 entirely
-	ContractTerminations []*OpenContract // Contracts that terminated this block.
-	MissedStorageProofs  []MissedStorageProof
-	SuccessfulWindows    []ContractID
-}
-
 // CreateGenesisState will create the state that contains the genesis block and
 // nothing else.
 func CreateGenesisState() (s *State, diffs []OutputDiff) {
